app_mockito/src/main/java/org/cbaron/appmockito/ejemplos/services/ExamenServiceImpl.java
package org.cbaron.appmockito.ejemplos.services;

import org.cbaron.appmockito.ejemplos.models.Examen;
import org.cbaron.appmockito.ejemplos.repositories.ExamenRepository;

import java.util.Optional;

public class ExamenServiceImpl implements ExamenService{

    private ExamenRepository examenRepository;

    public ExamenServiceImpl(ExamenRepository examenRepository) {
        this.examenRepository = examenRepository;
    }

    @Override
    public Optional<Examen> findExamenPorNumber(String nombre) {
        return examenRepository.findAll()
                .stream()
                .filter(examen -> examen.getNombre()
                .contains(nombre))
                .findFirst();
    }
}
